+++ methods/src/test/kotlin/methods/equationsystems/SolveByEliminationTest.kt
@@ -439,9 +439,14 @@ class SolveByEliminationTest {
 
                     step {
                         fromExpr = "2 x = 2"
+                        toExpr = "x = 1"
+                    }
+
+                    step {
+                        fromExpr = "x = 1"
                         toExpr = "SetSolution[x: {1}]"
                         explanation {
-                            key = EquationsExplanation.SolveLinearEquation
+                            key = EquationsExplanation.ExtractSolutionFromEquationInSolvedForm
                         }
                     }
                 }